libbinder: finish{F,Unf}lattenBinder symmetry
Before, both of these functions wrote stability information, but only
finishFlattenBinder wrote binder-specific objects to the Parcel. This
was because in order to function, unflattenBinder needed to read object
data, but flattenBinder was able to delegate this. Now, the
binder-specific object logic is moved into flattenBinder. This
organization makes it easy to re-use finishFlattenBinder for RPC
binders, which would write stability information but not the
binder-driver-specific objects.

Bug: 167966510
Test: boot
Change-Id: I0d1e48bf439dfd928e284c7c8c4284cd955a2850
Signed-off-by: Akash <akashniki@gmail.com>
Signed-off-by: DennySPb <dennyspb@gmail.com>